Move "SourceParams" to separate module
"SourceParams" and "_source_params_descriptions" were becoming very lengthy, therefore:
1) Introduce new module "utility/sourceparams.py"
2) Add "SourceParams", "_source_params_descriptions" and "_file_fields" there.
3) Also move the "assert all(p.value in _source_params_descriptions for p in SourceParams)" check there.
4) Add a new function "make_measurements_dataframe", for collecting multiple Numpy ndarrays into a single Pandas df, to the new module. Its use will become clear in an upcoming commit.
